+++ src/main/java/dev/dto/ConditionnementDtoResponse.java
@@ -14,7 +14,7 @@ public class ConditionnementDtoResponse {
 	private Integer id;
 	private Emballage emballage;
 	private Integer nombrePiece;
-	private BigDecimal poidsPiece;
+	private Integer poidsPiece;
 	private Unite unite;
 
 	public Integer getId() {
@@ -35,10 +35,10 @@ public class ConditionnementDtoResponse {
 	public void setNombrePiece(Integer nombrePiece) {
 		this.nombrePiece = nombrePiece;
 	}
-	public BigDecimal getPoidsPiece() {
+	public Integer getPoidsPiece() {
 		return poidsPiece;
 	}
-	public void setPoidsPiece(BigDecimal poidsPiece) {
+	public void setPoidsPiece(Integer poidsPiece) {
 		this.poidsPiece = poidsPiece;
 	}
 	public Unite getUnite() {
